ServerSide/requirements.txt
# Core PDF Processing Dependencies
PyMuPDF==1.23.5
numpy>=1.21.0
pytesseract>=0.3.8
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=6.0

# NLP and Text Processing Dependencies
scikit-learn>=1.0.0
nltk>=3.7
sentence-transformers>=2.0.0
transformers>=4.20.0

# Web Framework Dependencies
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Additional utilities
regex>=2022.1.0
tqdm>=4.60.0
aiofiles>=0.8.0
